moved openInBrowser to utils
